+++ src/forms/output/chartFormatting.js
